sketch for f.p.s./ mmorpg games
This is a simple command prompt file maker/reader, it was/is intended for creating a basic sketch for a game. It's primary use is for simple details on weapons/players and it's quite helpful, because whenever I have an idea for one i just pop this open and enter it in.